devops/requirements/requirements.txt
# common
awscli==1.11.125
boto3==1.4.4
Flask==0.10.1
kafka-python==1.3.1
PyJWT==1.4.2
python-dateutil==2.6.0
redis==2.10.3
requests==2.10.0
retrying==1.3.3
voluptuous==0.9.3
PyJWT==1.4.2
gevent==1.2.0
# Jinja2 needs to be pinned due to a problem with Jinja 2.9 (SyntaxError: 'yield' inside async function)
Jinja2==2.8
future==0.15.2
google-cloud==0.26.1

# axagent
GitPython==1.0.2
jira==1.0.7
flower==0.9.1

# axscheduler
APScheduler==3.2.0

# fixturemanager
pymongo==3.4.0

# gateway
django==1.9
django-extensions==1.6.7
django-filter==0.13.0
djangorestframework==3.3.3
django-crontab==0.7.1
requests==2.10.0
uwsgi==2.0.14
GitPython==1.0.2
pika==0.10.0
redlock-py==1.0.8

# axworkflowadc
transitions==0.4.0

# chaosmonkey
urllib3==1.18
websocket-client==0.37.0
PyYAML==3.12
certifi==2016.9.26

# test
pytest==3.0.1
responses==0.5.1
